+++ app/javascript/components/Home.jsx
@@ -1,6 +1,6 @@
 // import * as React from 'react';
 import React, { useEffect, useState } from 'react'
-import Universities from './universities/Universities'
+import HomePage from '../pages/HomePage'
 // import University from './university/University';
 // const cards = [1, 2, 3, 4, 5, 6, 7, 8, 9];
 
@@ -10,7 +10,7 @@ const Home = (props) => {
   console.log(props)
   return (
     <div>
-      <Universities/>
+      <HomePage/>
     </div>
   )
 }
