--- app/javascript/components/test.jsx
@@ -1,3 +1,0 @@
-document.addEventListener('DOMContentLoaded', () => {
-
-})
