+++ app/javascript/packs/application.js
@@ -24,13 +24,11 @@ import "bootstrap";
 // Internal imports, e.g:
 // import { initSelect2 } from '../components/init_select2';
 
-// import ShowModules from "../components/show_module";
 import "./Index"
 
 
 document.addEventListener('turbolinks:load', () => {
   // Call your functions here, e.g:
   // initSelect2();
-  // ShowModules();
   // console.log("test")
 });
