+++ app/views/pages/home.html.erb
@@ -2,12 +2,11 @@
 <%# <p>Find me in app/views/pages/home.html.erb</p> %>
 
 
-<%= render 'shared/intro' %>
+<%#= render 'shared/intro' %>
 
 <%# <a href="universities/1">test</a>
 <%= link_to "test", "universities/1" %>
-<%= link_to "Login", new_user_session_path, class: "nav-link" %>
-<%= link_to "Log out", destroy_user_session_path, method: :delete %>
+
 <div id="root"></div>
 <%# <div id="root2"></div> %>
 <%# <div id="Universities"></div> %>
